+++ lib/pycall/pycall_truffleruby.rb
@@ -1,0 +1,3 @@
+module PyCall
+
+end
